Allow pow to accept float exponents

+++ calculator.py
@@ -38,7 +38,8 @@ while calculation != "q":
 		print cube(int(token[1]))
 
 	elif token[0] == "pow":
-		print power(int(token[1]), int(token[2]))
+		#test if user inputs a float
+		print power(float(token[1]), float(token[2]))
 
 	elif token[0] == "mod":
 		print mod(int(token[1]), int(token[2]))
